Location as interface #1053
